+++ Fracciones.cs
@@ -19,11 +19,11 @@ class Fracciones:Calcular{
     {
         return base.concepto;
     }
-    public override float alCuadrado()
+    public override double alCuadrado()
     {
         return base.concepto;
     }
-    public override float potencia()
+    public override double potencia()
     {
         return base.concepto;
     }
